Moved token exceptions into security token folder
