+++ src/main/java/com/streisky/virtualshowcasebackend/config/security/token/TokenService.java
@@ -9,8 +9,8 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.JWTCreationException;
 import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.streisky.virtualshowcasebackend.entity.account.Account;
-import com.streisky.virtualshowcasebackend.exception.TokenCreationFailedException;
-import com.streisky.virtualshowcasebackend.exception.TokenInvalidException;
+import com.streisky.virtualshowcasebackend.config.security.token.exception.TokenCreationFailedException;
+import com.streisky.virtualshowcasebackend.config.security.token.exception.TokenInvalidException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
