+++ src/main/java/com/streisky/virtualshowcasebackend/config/security/token/exception/TokenInvalidException.java
@@ -1,4 +1,4 @@
-package com.streisky.virtualshowcasebackend.exception;
+package com.streisky.virtualshowcasebackend.config.security.token.exception;
 
 public class TokenInvalidException extends RuntimeException {
